chore: rename falcon module back to rpo_falcon512
