+++ src/dsa/mod.rs
@@ -1,3 +1,3 @@
 //! Digital signature schemes supported by default in the Miden VM.
 
-pub mod falcon512;
+pub mod rpo_falcon512;
